move getATech out of ButtonHelper BH is now under 350kb! Can be indexed
